+++ config/di-web.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 use Yiisoft\EventDispatcher\Provider\ListenerCollection;
 use Yiisoft\Yii\Event\ListenerCollectionFactory as Factory;
+use Yiisoft\Config\Config;
 
-/** @var \Yiisoft\Config\Config $config */
+/** @var Config $config */
 /** @var array $params */
 
 return [
-    ListenerCollection::class => static fn (Factory $factory) => $factory->create($config->get($params['yiisoft/yii-event']['eventsConfigGroup'])),
+    ListenerCollection::class => static fn(Factory $factory) => $factory->create($config->get($params['yiisoft/yii-event']['eventsConfigGroup'])),
 ];