+++ src/CallableFactory.php
@@ -19,9 +19,8 @@ use function is_string;
 final class CallableFactory
 {
     public function __construct(
-        private readonly ContainerInterface $container
-    ) {
-    }
+        private readonly ContainerInterface $container,
+    ) {}
 
     /**
      * Create real callable listener from definition.